post, report, like, comment update

+++ src/application/posts/services.py
@@ -20,11 +20,12 @@ class PostAppService:
         self.db_session = session
         self.post_service = PostService(session=self.db_session)
 
-    def get_post_by_id(self, id: uuid.UUID) -> Post:
+    def get_post_by_id(self, id: uuid.UUID) -> Optional[Post]:
         """get post using post id"""
         post = self.post_service.get_post_by_id(id=id)
         if not post:
-            raise NotFoundException(get_post_not_found())
+            # raise NotFoundException(get_post_not_found())
+            return None
         return post
 
     def get_all_posts_by_user_id(self, user_id: uuid.UUID) -> List[Post]:
@@ -42,16 +43,17 @@ class PostAppService:
 
     def get_post_by_post_id_for_user(
         self, post_id: uuid.UUID, user_id: uuid.UUID
-    ) -> Post:
+    ) -> Optional[Post]:
         """get post by post_id from user's post"""
         post = self.get_post_by_id(id=post_id)
         if not post or post.posted_by != user_id:
-            raise NotFoundException(get_post_not_found())
+            # raise NotFoundException(get_post_not_found())
+            return None
         return post
 
     def create_post(self, post: PostSchema) -> Post:
         """create new post"""
-        db_post = self.post_service.create_post(post=post)
+        db_post = self.post_service.create(post=post)
         return db_post
 
     def update_post(self, post_id: uuid.UUID, post: PostSchema) -> Post:
@@ -59,12 +61,16 @@ class PostAppService:
         db_post = self.get_post_by_post_id_for_user(
             post_id=post_id, user_id=post.posted_by
         )
-        return self.post_service.update_post(post=post, db_post=db_post)
+        if not db_post:
+            raise NotFoundException(get_post_not_found())
+        return self.post_service.update(post=post, db_post=db_post)
 
     def delete_post(self, post_id: uuid.UUID, user_id: uuid.UUID) -> None:
         """delete post"""
         db_post = self.get_post_by_post_id_for_user(post_id=post_id, user_id=user_id)
+        if not db_post:
+            return None
         # delete from minio
         for file in db_post.media:
             Boto3Service().delete_file(object_key=file.media_url)
-        self.post_service.delete_post(db_post=db_post)
+        self.post_service.delete(db_post=db_post)
